download and send pdf
- renamed the component
- merged the send pdf to email and download pdf
